chap08/pro08/src/main/java/org/mklinkj/taojwp/sec01/ex03/FirstServlet.java
package org.mklinkj.taojwp.sec01.ex03;

import static org.mklinkj.taojwp.common.Constants.HTML_CONTENT_TYPE;

import jakarta.servlet.annotation.WebServlet;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.io.PrintWriter;

@WebServlet("/first3")
public class FirstServlet extends HttpServlet {
  @Override
  protected void doGet(HttpServletRequest request, HttpServletResponse response)
      throws IOException {
    response.setContentType(HTML_CONTENT_TYPE);
    PrintWriter out = response.getWriter();
    out.print(
        """
        <script type='text/javascript'>
        location.href='second?forwardingType=location&name=lee';
        </script>
        """);
  }
}
